CR 6535301, Irbis: runtime performance regression
The regression was caused by missing VM hinting of startup phase finish. 
The problem was investigated and located by Andrei Pangin, it was 
intorduces during refactoring of the CLDC/CDC MIDlet suite loader code.

MIDlet startup phase is logically splitted into 2 parts:

1) system initializations of the enviroment to start MIDlet in
   (Isolate event queue, listeners, producers, security etc.)
2) MIDlet suite initialization (MIDlet state handler, constructor, etc.)

Each phase is notified to VM with start/end hints. The hinting of the end
of part 1) was lost during the refactoring, so VM continued to work in 
runtime with no JIT. That was one of the regression reasons.

Replicated from midp-feature-mr2-active RR branch, -r3467
Reviewed by Alexey Zavitaev, thanks.